PHP-155: Updating Windows build for PHP v7.1.x
* Updating dependency versions for Windows build
* Adding PHP v7.1.7 to AppVeyor matrix and re-ordering build
* Fixing a default casting issue for features when using PHP 7.1

+++ features/bootstrap/FeatureContext.php
@@ -74,7 +74,7 @@ class FeatureContext implements Context, SnippetAcceptingContext
     public function prepareTestFolders()
     {
         $dir = sys_get_temp_dir().DIRECTORY_SEPARATOR.'php-driver'.DIRECTORY_SEPARATOR.
-            md5(microtime() * rand(0, 10000));
+            md5((int)microtime() * rand(0, 10000));
         $phpFinder = new PhpExecutableFinder();
         if (false === $php = $phpFinder->find()) {
             throw new \RuntimeException('Unable to find the PHP executable.');
